Refactor migrations wip
