+++ migration/migrate.go
@@ -2,6 +2,7 @@ package migration
 
 import (
 	"encoding/json"
+	"errors"
 	liberr "github.com/konveyor/controller/pkg/error"
 	"github.com/konveyor/tackle2-hub/database"
 	"github.com/konveyor/tackle2-hub/model"
@@ -45,20 +46,29 @@ func Migrate(migrations []Migration) (err error) {
 		}
 	}
 
+	var start = v.Version
+	if start == 0 {
+		start = SupportedFrom
+	} else if start < SupportedFrom {
+		err = errors.New("unsupported database version")
+		log.Error(err, "Unable to migrate database.", "version", v.Version)
+		return
+	}
+
 	// Version is the index of the last successful migration,
 	// so we want to start iteration at the next index.
 	migrations = append([]Migration{nil}, migrations...)
-	for i := v.Version + 1; i < len(migrations); i++ {
+	for i := start + 1; i < len(migrations); i++ {
 		m := migrations[i]
 
 		db, err = database.Open(false)
 		if err != nil {
-			err = liberr.Wrap(err, "version", m.Name())
+			err = liberr.Wrap(err, "version", i)
 			return
 		}
 
 		f := func(db *gorm.DB) (err error) {
-			log.Info("Running migration.", "version", m.Name())
+			log.Info("Running migration.", "version", i)
 			err = m.Apply(db)
 			if err != nil {
 				return
@@ -71,13 +81,21 @@ func Migrate(migrations []Migration) (err error) {
 		}
 		err = db.Transaction(f)
 		if err != nil {
-			err = liberr.Wrap(err, "version", m.Name())
+			err = liberr.Wrap(err, "version", i)
 			return
 		}
 
 		err = database.Close(db)
 		if err != nil {
-			err = liberr.Wrap(err, "version", m.Name())
+			err = liberr.Wrap(err, "version", i)
+			return
+		}
+	}
+
+	if Settings.Hub.Development {
+		log.Info("Running development auto-migration.")
+		err = autoMigrate(db, migrations[len(migrations)-1].Models())
+		if err != nil {
 			return
 		}
 	}
@@ -95,6 +113,27 @@ func setVersion(db *gorm.DB, version int) (err error) {
 	result := db.Where("key", VersionKey).Updates(setting)
 	if result.Error != nil {
 		err = liberr.Wrap(result.Error)
+		return
+	}
+	return
+}
+
+//
+// AutoMigrate the database.
+func autoMigrate(db *gorm.DB, models []interface{}) (err error) {
+	db, err = database.Open(false)
+	if err != nil {
+		err = liberr.Wrap(err)
+		return
+	}
+	err = db.AutoMigrate(models)
+	if err != nil {
+		err = liberr.Wrap(err)
+		return
+	}
+	err = database.Close(db)
+	if err != nil {
+		err = liberr.Wrap(err)
 		return
 	}
 	return
