+++ migration/pkg.go
@@ -2,17 +2,23 @@ package migration
 
 import (
 	"github.com/konveyor/controller/pkg/logging"
-	"github.com/konveyor/tackle2-hub/migration/v1"
 	"github.com/konveyor/tackle2-hub/migration/v2"
 	"github.com/konveyor/tackle2-hub/migration/v3"
+	"github.com/konveyor/tackle2-hub/settings"
 	"gorm.io/gorm"
 )
 
 var log = logging.WithName("migration")
+var Settings = &settings.Settings
 
 //
 // VersionKey is the setting containing the migration version.
 const VersionKey = ".migration.version"
+
+//
+// SupportedFrom is the index of the
+// earliest version that we can migrate from.
+const SupportedFrom = 1
 
 //
 // Version represents the value of the .migration.version setting.
@@ -24,14 +30,14 @@ type Version struct {
 // Migration encapsulates the functionality necessary to perform a migration.
 type Migration interface {
 	Apply(*gorm.DB) error
-	Name() string
+	Models() []interface{}
 }
 
 //
 // All migrations in order.
 func All() []Migration {
 	return []Migration{
-		v1.Migration{},
+		nil, // v1
 		v2.Migration{},
 		v3.Migration{},
 	}
