--- migration/v1/model/bucket.go
@@ -1,5 +1,0 @@
-package model
-
-type BucketOwner struct {
-	Bucket string `gorm:"index"`
-}
