+++ migration/v2/model/importsummary.go
@@ -1,0 +1,10 @@
+package model
+
+type ImportSummary struct {
+	Model
+	Content        []byte
+	Filename       string
+	ImportStatus   string
+	Imports        []Import `gorm:"constraint:OnDelete:CASCADE"`
+	CreateEntities bool
+}
