+++ migration/v2/model/importtag.go
@@ -1,0 +1,9 @@
+package model
+
+type ImportTag struct {
+	Model
+	Name     string
+	TagType  string
+	ImportID uint `gorm:"index"`
+	Import   *Import
+}
