+++ migration/v2/model/jobfunction.go
@@ -1,0 +1,8 @@
+package model
+
+type JobFunction struct {
+	Model
+	Username     string
+	Name         string        `gorm:"index;unique;not null"`
+	Stakeholders []Stakeholder `gorm:"constraint:OnDelete:SET NULL"`
+}
