+++ migration/v2/model/stakeholder.go
@@ -1,0 +1,11 @@
+package model
+
+type Stakeholder struct {
+	Model
+	Name             string             `gorm:"not null;"`
+	Email            string             `gorm:"index;unique;not null"`
+	Groups           []StakeholderGroup `gorm:"many2many:StakeholderGroupStakeholder;constraint:OnDelete:CASCADE"`
+	BusinessServices []BusinessService  `gorm:"constraint:OnDelete:SET NULL"`
+	JobFunctionID    *uint              `gorm:"index"`
+	JobFunction      *JobFunction
+}
