+++ migration/v2/model/stakeholdergroup.go
@@ -1,0 +1,9 @@
+package model
+
+type StakeholderGroup struct {
+	Model
+	Name         string `gorm:"index;unique;not null"`
+	Username     string
+	Description  string
+	Stakeholders []Stakeholder `gorm:"many2many:StakeholderGroupStakeholder;constraint:OnDelete:CASCADE"`
+}
