+++ migration/v2/model/tag.go
@@ -1,0 +1,9 @@
+package model
+
+type Tag struct {
+	Model
+	Name      string `gorm:"uniqueIndex:tagA;not null"`
+	Username  string
+	TagTypeID uint `gorm:"uniqueIndex:tagA;index;not null"`
+	TagType   TagType
+}
