+++ migration/v3/model/ticket.go
@@ -2,20 +2,6 @@ package model
 
 import "time"
 
-type Tracker struct {
-	Model
-	Name        string `gorm:"index;unique;not null"`
-	URL         string
-	Kind        string
-	Identity    *Identity
-	IdentityID  uint
-	Metadata    JSON
-	Connected   bool
-	LastUpdated time.Time
-	Message     string
-	Tickets     []Ticket `gorm:"constraint:OnDelete:CASCADE"`
-}
-
 type Ticket struct {
 	Model
 	// Kind of ticket in the external tracker.
